COMP: Fixup compilations warning (lost of precision)
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=16830
git-svn-id: http://svn.slicer.org/Slicer4/trunk@16830 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Libs/MRML/vtkMRMLDisplayNode.h
@@ -234,7 +234,7 @@ class VTK_MRML_EXPORT vtkMRMLDisplayNode : public vtkMRMLNode
   /// If 0, display in all views
   int GetNumberOfViewNodeIDs()const
   {
-    return ViewNodeIDs.size();
+    return static_cast<int>(ViewNodeIDs.size());
   }
 
   /// Get View Node ID's for the view to display this node in.
